+++ qualang_tools/config/primitive_components.py
@@ -1,10 +1,28 @@
-from typing import Any, Dict, List
+import abc
+from typing import Any, Dict, List, Optional
 
 
-class Port(object):
+class ConfigBuilderClass(abc.ABC):
+    def __init__(self, **kwargs) -> None:
+        super().__init__()
+
+
+class ConfigBuilderElement(ConfigBuilderClass):
+    def __init__(self, name, **kwargs) -> None:
+        super(ConfigBuilderElement, self).__init__()
+        self.name = name
+
+
+class Port(ConfigBuilderClass):
     def __init__(self, controller: str, port_id: int, offset: float = 0) -> None:
-        self.info = (controller, port_id)
+        super(Port, self).__init__()
+        self.controller = controller
+        self.port_id = port_id
         self.offset = offset
+
+    @property
+    def info(self):
+        return self.controller, self.port_id
 
     def __eq__(self, other) -> bool:
         return self.info == other.info and type(self) == type(other)
@@ -24,33 +42,55 @@ class Port(object):
 
 
 class AnalogInputPort(Port):
-    def __init__(self, controller: str, port_id: int, offset: float = 0):
-        self.gain_db = None
-        super().__init__(controller, port_id, offset)
+    def __init__(
+        self,
+        controller: str,
+        port_id: int,
+        offset: float = 0,
+        gain_db: Optional[float] = None,
+    ):
+        super(AnalogInputPort, self).__init__(controller, port_id, offset)
+        self.gain_db = gain_db
 
 
 class AnalogOutputPort(Port):
-    def __init__(self, controller: str, port_id: int, offset: float = 0):
-        self.delay = None
-        self.filter = None
-        self.channel_weights = None
-        super().__init__(controller, port_id, offset)
+    def __init__(
+        self,
+        controller: str,
+        port_id: int,
+        offset: float = 0,
+        delay: Optional[float] = None,
+        filter: Optional[float] = None,
+        channel_weights: Optional[float] = None,
+    ):
+        super(AnalogOutputPort, self).__init__(controller, port_id, offset)
+        self.delay = delay
+        self.filter = filter
+        self.channel_weights = channel_weights
 
 
 class DigitalInputPort(Port):
-    def __init__(self, controller: str, port_id: int, offset: float = 0):
-        self.polarity = None
-        self.window = None
-        self.threshold = None
-        super().__init__(controller, port_id, offset)
+    def __init__(
+        self,
+        controller: str,
+        port_id: int,
+        offset: float = 0,
+        polarity: Optional[float] = None,
+        window: Optional[float] = None,
+        threshold: Optional[float] = None,
+    ):
+        super(DigitalInputPort, self).__init__(controller, port_id, offset)
+        self.polarity = polarity
+        self.window = window
+        self.threshold = threshold
 
 
 class DigitalOutputPort(Port):
     def __init__(self, controller: str, port_id: int, offset: float = 0):
-        super().__init__(controller, port_id, offset)
+        super(DigitalOutputPort, self).__init__(controller, port_id, offset)
 
 
-class Waveform(object):
+class Waveform(ConfigBuilderElement):
     def __init__(self, name: str, dictionary: Dict[str, Any]):
         """A waveform played by the QOP
 
@@ -59,12 +99,19 @@ class Waveform(object):
         :param dictionary: parameters to initialize a WF
         :type dictionary: Dict[str, Any]
         """
-        self.name = name
+        super(Waveform, self).__init__(name)
         self.dict = dictionary
 
 
-class Pulse(object):
-    def __init__(self, name: str, wfs: List[Waveform], operation: str, length: int):
+class Pulse(ConfigBuilderElement):
+    def __init__(
+        self,
+        name: str,
+        wfs: List[Waveform],
+        operation: str,
+        length: int,
+        digital_marker: Optional[float] = None,
+    ):
         """A an analog pulse
 
         :param name: Name for this pulse
@@ -76,9 +123,13 @@ class Pulse(object):
         :param length:
         :type length: int
         """
+        super(Pulse, self).__init__(name)
         assert len(wfs) <= 2
-        self.name = name
         self.wfs = wfs
+        self.operation = operation
+        self.length = length
+        self.digital_marker = digital_marker
+
         self.dict = dict()
         self.dict["operation"] = operation
         self.dict["length"] = length
@@ -88,20 +139,20 @@ class Pulse(object):
             self.dict["waveforms"]["Q"] = wfs[1].name
         elif len(wfs) == 1:
             self.dict["waveforms"]["single"] = wfs[0].name
-        self.digital_marker = None
 
     @property
     def waveform_names(self):
         return [wf.name for wf in self.wfs]
 
 
-class Operation(object):
+class Operation(ConfigBuilderClass):
     def __init__(self, pulse: Pulse, name: str = ""):
+        super(Operation, self).__init__()
         self.pulse = pulse
         self.name = pulse.name if name == "" else name
 
 
-class IntegrationWeights(object):
+class IntegrationWeights(ConfigBuilderElement):
     def __init__(self, name: str, cosine: List, sine: List):
         """A vector of integration weights used in integration and demodulation
 
@@ -112,19 +163,36 @@ class IntegrationWeights(object):
         :param sine: a list of weights to be used with the sine demodulation element, specified at 1 sample per nanosecond
         :type sine: List
         """
-        self.name = name
+        super(IntegrationWeights, self).__init__(name)
         self.dict = dict()
         self.dict["cosine"] = cosine
         self.dict["sine"] = sine
 
+    @property
+    def cosine(self):
+        return self.dict["cosine"]
 
-class Weights(object):
+    @cosine.setter
+    def cosine(self, w: List[float]):
+        self.dict["cosine"] = w
+
+    @property
+    def sine(self):
+        return self.dict["sine"]
+
+    @sine.setter
+    def sine(self, w: List[float]):
+        self.dict["sine"] = w
+
+
+class Weights(ConfigBuilderClass):
     def __init__(self, weights: IntegrationWeights, name: str = ""):
+        super(Weights, self).__init__()
         self.name = weights.name if name == "" else name
         self.weights = weights
 
 
-class DigitalSample(object):
+class DigitalSample(ConfigBuilderClass):
     def __init__(self, state: int, duration: int):
         """A sample to describe digital waveform
 
@@ -133,25 +201,27 @@ class DigitalSample(object):
         :param duration: duration of the output in nanoseconds
         :type duration: int
         """
+        super(DigitalSample, self).__init__()
         assert state == 0 or state == 1
         self.state = state
         self.duration = duration
 
 
-class Matrix2x2(object):
+class Matrix2x2(ConfigBuilderClass):
     def __init__(self, data: List[List[float]]):
         """A correction matrix for the IQ mixer
 
         :param data: 2x2 matrix provided as a list containing a list of rows
         :type data: List[List[float]]
         """
+        super(Matrix2x2, self).__init__()
         assert len(data) == 2
         assert len(data[0]) == 2
         assert len(data[1]) == 2
         self.data = data
 
 
-class AnalogOutputFilter(object):
+class AnalogOutputFilter(ConfigBuilderClass):
     def __init__(self, feedback: List[float], feedforward: List[float]):
         """A filter applied to the analog output ports
 
@@ -160,5 +230,6 @@ class AnalogOutputFilter(object):
         :param feedforward: feedforward taps for the output filter
         :type feedforward: List[float]
         """
+        super(AnalogOutputFilter, self).__init__()
         self.feedback = feedback
         self.feedforward = feedforward